requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5.7
azure-common==1.1.27
azure-core==1.19.0
azure-functions==1.7.2
azure-identity==1.6.1
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.8
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-comments-xtd==2.8.0
django-contrib-comments==1.9.2
django-cors-headers==3.10.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-filter==21.1
django-mathfilters==1.0.0
django-mssql==1.2
django-mssql-backend==2.8.1
django-multiselectfield==0.1.12
django-pyodbc==1.1.3
django-pyodbc-azure==1.2.5
django-redis-cache==2.1.0
django-rq==2.1.0
django-storages==1.11.1
django-storages-azure==1.6.8
django-summernote==0.8.11.6
django-unixtimestampfield==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
docutils==0.16
et-xmlfile==1.1.0
fontawesome-free==5.15.4
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.2
image==1.5.33
importlib-metadata==4.8.1
importlib-resources==5.4.0
isodate==0.6.0
jsonschema==4.0.0
kombu==5.1.0
Markdown==3.3.6
msal==1.14.0
msal-extensions==0.3.0
msrest==0.6.19
mssql-django==1.1.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==21.3
Pillow==8.3.1
polib==1.1.1
portalocker==1.7.1
prompt-toolkit==3.0.30
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyodbc==4.0.32
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.9.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
tzdata==2022.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.5.0
wrapt==1.14.1
zipp==3.6.0
pre-commit==3.5.0
ruff==0.1.8
